Prisma AIRS refactor (Option B) Phases 1-2: shared ApiModule + Model Security integration
Split the 150-command monolithic integration into a shared transport module
plus per-domain integrations in the same pack.

Phase 1 — Extract Packs/ApiModules/Scripts/PrismaAirsApiModule:
  Client(BaseClient) with OAuth2 client_credentials auth to SCM, http_request
  with the full plane-routing flag set, scanner_request, test_module, and
  shared constants. All credential/URL fields optional so scoped-credential
  integrations construct cleanly. 21 unit tests.

Phase 2 — Split out Integrations/PrismaAIRsModelSecurity:
  28 model_security_* commands moved verbatim from the monolith with a scoped
  config (url, credentials, tsg_id, insecure, proxy), 36 unit tests, README,
  description, and shared product image. Deleted the same commands, dispatch,
  yml defs, README sections, and tests from the monolith in this commit to
  avoid duplicate command names.

Counts reconcile: monolith 122 cmds/209 tests + Model Security 28 cmds/36 tests
= 150 cmds/245 original tests preserved, + 21 ApiModule tests = 266 total.
All three items pass demisto-sdk validate; verified live end-to-end by running
the Model Security test playbook against a new incident on the XSOAR server.
